Support sub tasks
Create dedicated task class and refactor the code in the report engine.
Add support for better ordering of tasks and sub tasks by using a tree
and not a list.
Create filter to format the date in template.
Replace colors in PlantUML by using HSL values and use variables for
readability.
Change methods scope to private when possible.